+++ OpenDiscussion/Controllers/DiscussionsController.cs
@@ -29,11 +29,13 @@ namespace OpenDiscussion.Controllers
             ViewBag.Discussions = discussions;
             return View();
         }
+        [Authorize(Roles = "User,Moderator,Admin")]
         public IActionResult Edit(int id)
         {
             Discussion discussion = db.Discussions.Find(id);
             return View(discussion);
         }
+        [Authorize(Roles = "User,Moderator,Admin")]
         [HttpPost]
         public IActionResult Edit(int id, Discussion requestDiscussion)
         {
@@ -41,10 +43,18 @@ namespace OpenDiscussion.Controllers
             requestDiscussion.DiscussionId = id;
             if (ModelState.IsValid)
             {
-                discussion.Title = requestDiscussion.Title;
-                discussion.Content = requestDiscussion.Content;
-                db.SaveChanges();
-                return RedirectToAction("Index", new {id = discussion.TopicId});
+                if (_userManager.GetUserId(User) == discussion.UserId || User.IsInRole("Moderator") || User.IsInRole("Admin"))
+                {
+                    discussion.Title = requestDiscussion.Title;
+                    discussion.Content = requestDiscussion.Content;
+                    db.SaveChanges();
+                    return RedirectToAction("Index", new { id = discussion.TopicId });
+                }
+                else
+                {
+                    TempData["MessageDeny"] = "Nu aveti voie sa editati o discutie care nu va apartine";
+                    return RedirectToAction("Index", new { id = discussion.TopicId }); 
+                }
             }
             else
                 return View(requestDiscussion);
@@ -75,34 +85,54 @@ namespace OpenDiscussion.Controllers
                 return View(discussion);
             }
         }
+        [Authorize(Roles = "User,Moderator,Admin")]
+        [HttpPost]
         public IActionResult Delete(int id)
         {
             Discussion discussion = db.Discussions.Include("Comments").Where(diss => diss.DiscussionId == id).First();
-            db.Discussions.Remove(discussion);
-            db.SaveChanges();
-            return RedirectToAction("Index", new { id = discussion.TopicId});
+            if (_userManager.GetUserId(User) == discussion.UserId || User.IsInRole("Moderator") || User.IsInRole("Admin"))
+            {
+                db.Discussions.Remove(discussion);
+                db.SaveChanges();
+                return RedirectToAction("Index", new { id = discussion.TopicId });
+            }
+            else
+            {
+                TempData["MessageDeny"] = "Nu aveti voie sa stergeti o discutie care nu va apartine";
+                return RedirectToAction("Index", new { id = discussion.TopicId });
+            }
+
         }
         [Authorize(Roles = "User,Moderator,Admin")]
         public IActionResult Show(int id)
         {
-            Discussion discussion = db.Discussions.Include("Comments").Include("User").Where(disc => disc.DiscussionId == id).First();
+            Discussion discussion = db.Discussions
+                                    .Include("Comments")
+                                    .Include("User")
+                                    .Include("Comments.User")
+                                    .Where(disc => disc.DiscussionId == id)
+                                    .First();
             return View(discussion);
         }
         [Authorize(Roles = "User,Moderator,Admin")]
         [HttpPost]
         public IActionResult Show([FromForm] Comment comment)
         {
+            comment.UserId = _userManager.GetUserId(User);
+            comment.Date = DateTime.Now;
             if (ModelState.IsValid)
-            {
-                comment.Date = DateTime.Now;
-
+            { 
                 db.Comments.Add(comment);
                 db.SaveChanges();
                 return RedirectToAction("Show", new { id = comment.DiscussionId });
             }
             else
             {
-                Discussion discussion = db.Discussions.Include("Comments").Include("User").Where(disc => disc.DiscussionId == comment.DiscussionId).First();
+                Discussion discussion = db.Discussions.Include("Comments")
+                                                      .Include("User")
+                                                      .Include("Comments.User")
+                                                      .Where(disc => disc.DiscussionId == comment.DiscussionId)
+                                                      .First();
                 return View(discussion);
             }
         }
